Refactor: Update order notification mechanism for unpaid orders and update tests

+++ service-api/src/main/java/greencity/service/ubs/NotificationService.java
@@ -242,5 +242,5 @@ public interface NotificationService {
      *
      * @author Vladyslav Haliara
      */
-    void notifyUnpaidOrderPermanently(Order order, Double sumToPay);
+    void notifyUnpaidOrderPermanently(Order order, Long sumToPay);
 }
